+++ scripts/devinfra/install-pwsh.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+if [[ ! $distribution = "ubuntu" ]]; then
+  echo "invalid distribution"
+  exit 1
+fi
 wget -q https://packages.microsoft.com/config/ubuntu/22.04/packages-microsoft-prod.deb
 sudo dpkg -i packages-microsoft-prod.deb
 rm -f packages-microsoft-prod.deb
