+++ scripts/k8stools/install-kubectx.sh
@@ -1,9 +1,21 @@
+#!/bin/bash
 git clone https://github.com/ahmetb/kubectx.git ~/.kubectx
 sudo mv ~/.kubectx/kubectx /usr/local/bin
 sudo chmod +x /usr/local/bin/kubectx
 sudo mv ~/.kubectx/kubens /usr/local/bin
 sudo chmod +x /usr/local/bin/kubens
-mkdir -p ~/.oh-my-zsh/completions
-chmod -R 755 ~/.oh-my-zsh/completions
-ln -s ~/.kubectx/completion/kubectx.zsh ~/.oh-my-zsh/completions/_kubectx.zsh
-ln -s ~/.kubectx/completion/kubens.zsh ~/.oh-my-zsh/completions/_kubens.zsh
+
+# kubectl bash completion
+sudo ln -s ~/.kubectx/completion/kubens.bash /etc/bash_completion.d/kubens
+sudo ln -s ~/.kubectx/completion/kubectx.bash /etc/bash_completion.d/kubectx
+
+# kubectl zsh completion
+if [ -e "${USERHOME}}/.oh-my-zsh" ]; then
+		mkdir -p "${USERHOME}/.oh-my-zsh/completions"
+		ln -s ~/.kubectx/completion/kubens.zsh ~/.oh-my-zsh/completions/_kubens.zsh
+		ln -s ~/.kubectx/completion/kubectx.zsh ~/.oh-my-zsh/completions/_kubectx.zsh
+		chown -R "${USERNAME}" "${USERHOME}/.oh-my-zsh"
+fi
+
+
+
